fix: fix relationship between the discount price and the total amount
